Give some love to tests
