+++ jest.config.js
@@ -6,6 +6,7 @@ module.exports = {
     globals: {
         'ts-jest': {
             tsConfig: 'tsconfig.json',
+            packageJson: 'package.json',
         },
     },
     testPathIgnorePatterns: ['.+\\.ignore\\.+'],
