+++ src/__tests__/customHook.ignore.ts
@@ -2,10 +2,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-module.exports = {
-    hooks: {
-        preoutput: () => {
-            return { test: 'test' };
-        },
+export const hooks = {
+    preoutput: () => {
+        return { test: 'test' };
     },
 };
