+++ src/hooks/datadog/__tests__/index.test.ts
@@ -2,31 +2,18 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
+import { mockStats, mockReport } from '../../../__tests__/testHelpers.ignore';
+
 describe('Datadog Hook', () => {
     const buildPluginMock = {
         options: {},
-    };
-    const statsMock = {
-        modules: [],
-        chunks: [],
-        assets: [],
-        warnings: [],
-        errors: [],
-        entrypoints: {},
-    };
-    const reportMock = {
-        timings: {
-            tapables: {},
-            loaders: {},
-        },
-        dependencies: {},
     };
 
     test('It should not fail given undefined options', async () => {
         const { hooks } = require('../index');
         const obj = await hooks.preoutput.call(buildPluginMock, {
-            report: reportMock,
-            stats: { toJson: () => statsMock },
+            report: mockReport,
+            stats: mockStats,
         });
 
         expect(typeof obj).toBe('object');
